requirements.txt
spotdl
customtkinter
python-dotenv
openai
Pillow
